+++ services/chat/src/common/paginated-response-with-total.dto.ts
@@ -1,0 +1,7 @@
+import { PaginatedResponse } from './paginated-response.dto';
+
+export interface PaginatedResponseWithTotalCount<
+  T,
+> extends PaginatedResponse<T> {
+  totalItems: number;
+}