+++ services/chat/src/conversations/dto/conversation-list-item.dto.ts
@@ -1,0 +1,21 @@
+import { ConversationID, UserID } from '../types';
+
+export interface ConversationListItemDTO {
+  id: ConversationID;
+  organization: {
+    id: UserID;
+    name: string;
+    avatar: string;
+  };
+  member: {
+    id: UserID;
+    name: string;
+    avatar: string;
+  };
+  lastMessage: {
+    content: string;
+    senderId: UserID;
+    timestamp: Date;
+  };
+  unreadCount: number;
+}